moved types.pyx to internals folder
