+++ startrak/__init__.py
@@ -1,6 +1,6 @@
 import astropy.io
-from . import types
-from . import io
+from .internals import types
+from . import io, sessions
 
 APPNAME = "StarTrak"
 VERSION = "1.0.0"
