+++ startrak/io.pyx
@@ -1,7 +1,7 @@
 import os
 import numpy as np
 from astropy.io import fits
-from startrak.types import FileInfo
+from startrak.internals.types import FileInfo
 
 
 # ----- Wrapper functions around astropy.io --------
